requirements.txt
fastapi
uvicorn
python-telegram-bot
pillow
